Feature: Refactor code
