+++ apps/web/lib/errors.ts
@@ -1,0 +1,11 @@
+export class FetchError extends Error {
+  status: number;
+  info?: unknown;
+
+  constructor(message: string, status: number, info?: unknown) {
+    super(message);
+    this.name = "FetchError";
+    this.status = status;
+    this.info = info;
+  }
+}
